snip links from genericapiserver to api/validation
